Library: Refactor entities
