+++ app/src/main/java/com/fingerprintjs/android/fpjs_pro_demo/base/BaseActivity.kt
