+++ app/src/main/res/values/strings.xml
@@ -8,8 +8,8 @@
     <string name="prefs_endpoint_url_key">endpointUrl</string>
     <string name="prefs_public_token_key">publicToken</string>
     <string name="ip_title">IP:</string>
-    <string name="os_title">OS</string>
-    <string name="location_title">Location</string>
+    <string name="os_title">OS:</string>
+    <string name="location_title">Location:</string>
     <string name="fingerprintjs">FingerprintJS</string>
     <string name="us_endpoint">🇺🇸 United States</string>
     <string name="us_endpoint_url">https://api.fpjs.io</string>
